NIFI-7342: Refactored NiFi SQL table models into their own JAR
