+++ nifi-nar-bundles/nifi-sql-reporting-bundle/nifi-sql-query-api/src/main/java/org/apache/nifi/reporting/sql/QueryResult.java
