+++ nifi-nar-bundles/nifi-sql-reporting-bundle/nifi-sql-query-api/src/main/java/org/apache/nifi/reporting/sql/metrics/JvmMetricsTable.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.nifi.reporting.sql.connectionstatus;
+package org.apache.nifi.reporting.sql.metrics;
 
 import org.apache.calcite.linq4j.AbstractEnumerable;
 import org.apache.calcite.linq4j.Enumerable;
@@ -34,42 +34,49 @@ import org.apache.calcite.schema.TranslatableTable;
 import org.apache.calcite.schema.impl.AbstractTable;
 import org.apache.calcite.util.Pair;
 import org.apache.nifi.logging.ComponentLog;
-import org.apache.nifi.reporting.ReportingContext;
+import org.apache.nifi.metrics.jvm.JmxJvmMetrics;
+import org.apache.nifi.metrics.jvm.JvmMetrics;
+import org.apache.nifi.reporting.util.metrics.MetricNames;
+import org.apache.nifi.reporting.util.metrics.MetricsService;
 
 import java.lang.reflect.Type;
-import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 
-public class ConnectionStatusTable extends AbstractTable implements QueryableTable, TranslatableTable {
+public class JvmMetricsTable extends AbstractTable implements QueryableTable, TranslatableTable {
 
     private final ComponentLog logger;
 
     private RelDataType relDataType = null;
 
-    private volatile ReportingContext context;
     private volatile int maxRecordsRead;
+    private volatile MetricsService metricsService = new MetricsService();
+    private volatile JvmMetrics virtualMachineMetrics;
 
-    private final Set<ConnectionStatusEnumerator> enumerators = new HashSet<>();
+    private final Set<JvmMetricsEnumerator> enumerators = new HashSet<>();
 
     /**
-     * Creates a Connection Status table.
+     * Creates a JVM Metrics table.
      */
-    public ConnectionStatusTable(final ReportingContext context, final ComponentLog logger) {
-        this.context = context;
+    public JvmMetricsTable(final ComponentLog logger) {
         this.logger = logger;
+        virtualMachineMetrics = JmxJvmMetrics.getInstance();
     }
 
     @Override
     public String toString() {
-        return "ConnectionStatusTable";
+        return "JvmMetricsTable";
     }
 
     public void close() {
         synchronized (enumerators) {
-            for (final ConnectionStatusEnumerator enumerator : enumerators) {
+            for (final JvmMetricsEnumerator enumerator : enumerators) {
                 enumerator.close();
             }
         }
@@ -86,7 +93,7 @@ public class ConnectionStatusTable extends AbstractTable implements QueryableTab
             @Override
             @SuppressWarnings({"unchecked", "rawtypes"})
             public Enumerator<Object> enumerator() {
-                final ConnectionStatusEnumerator connectionStatusEnumerator = new ConnectionStatusEnumerator(context, logger, fields) {
+                final JvmMetricsEnumerator jvmMetricsEnumerator = new JvmMetricsEnumerator(logger, fields) {
                     @Override
                     protected void onFinish() {
                         final int recordCount = getRecordsRead();
@@ -105,10 +112,10 @@ public class ConnectionStatusTable extends AbstractTable implements QueryableTab
                 };
 
                 synchronized (enumerators) {
-                    enumerators.add(connectionStatusEnumerator);
+                    enumerators.add(jvmMetricsEnumerator);
                 }
 
-                return connectionStatusEnumerator;
+                return jvmMetricsEnumerator;
             }
         };
     }
@@ -142,7 +149,7 @@ public class ConnectionStatusTable extends AbstractTable implements QueryableTab
             fields[i] = i;
         }
 
-        return new ConnectionStatusTableScan(context.getCluster(), relOptTable, this, fields);
+        return new JvmMetricsTableScan(context.getCluster(), relOptTable, this, fields);
     }
 
     @Override
@@ -151,48 +158,41 @@ public class ConnectionStatusTable extends AbstractTable implements QueryableTab
             return relDataType;
         }
 
-        final List<String> names = Arrays.asList(
-                "id",
-                "groupId",
-                "name",
-                "sourceId",
-                "sourceName",
-                "destinationId",
-                "destinationName",
-                "backPressureDataSizeThreshold",
-                "backPressureBytesThreshold",
-                "backPressureObjectThreshold",
-                "isBackPressureEnabled",
-                "inputCount",
-                "inputBytes",
-                "queuedCount",
-                "queuedBytes",
-                "outputCount",
-                "outputBytes",
-                "maxQueuedCount",
-                "maxQueuedBytes"
-        );
-        final List<RelDataType> types = Arrays.asList(
-                typeFactory.createJavaType(String.class),
-                typeFactory.createJavaType(String.class),
-                typeFactory.createJavaType(String.class),
-                typeFactory.createJavaType(String.class),
-                typeFactory.createJavaType(String.class),
-                typeFactory.createJavaType(String.class),
-                typeFactory.createJavaType(String.class),
-                typeFactory.createJavaType(String.class),
-                typeFactory.createJavaType(long.class),
-                typeFactory.createJavaType(long.class),
-                typeFactory.createJavaType(boolean.class),
-                typeFactory.createJavaType(int.class),
-                typeFactory.createJavaType(long.class),
-                typeFactory.createJavaType(int.class),
-                typeFactory.createJavaType(long.class),
-                typeFactory.createJavaType(int.class),
-                typeFactory.createJavaType(long.class),
-                typeFactory.createJavaType(int.class),
-                typeFactory.createJavaType(long.class)
-        );
+        final List<String> names = Stream.of(
+                MetricNames.JVM_DAEMON_THREAD_COUNT,
+                MetricNames.JVM_THREAD_COUNT,
+                MetricNames.JVM_THREAD_STATES_BLOCKED,
+                MetricNames.JVM_THREAD_STATES_RUNNABLE,
+                MetricNames.JVM_THREAD_STATES_TERMINATED,
+                MetricNames.JVM_THREAD_STATES_TIMED_WAITING,
+                MetricNames.JVM_UPTIME,
+                MetricNames.JVM_HEAP_USED,
+                MetricNames.JVM_HEAP_USAGE,
+                MetricNames.JVM_NON_HEAP_USAGE,
+                MetricNames.JVM_FILE_DESCRIPTOR_USAGE
+        ).map((name) -> name.replace(".", "_").replace("-","_")).collect(Collectors.toList());
+        final List<RelDataType> types = new ArrayList<>();
+        types.add(typeFactory.createJavaType(int.class));
+        types.add(typeFactory.createJavaType(int.class));
+        types.add(typeFactory.createJavaType(int.class));
+        types.add(typeFactory.createJavaType(int.class));
+        types.add(typeFactory.createJavaType(int.class));
+        types.add(typeFactory.createJavaType(int.class));
+        types.add(typeFactory.createJavaType(long.class));
+        types.add(typeFactory.createJavaType(double.class));
+        types.add(typeFactory.createJavaType(double.class));
+        types.add(typeFactory.createJavaType(double.class));
+        types.add(typeFactory.createJavaType(double.class));
+
+        // Add fields for the garbage collectors
+        metricsService.getMetrics(virtualMachineMetrics);
+        for (Map.Entry<String, JvmMetrics.GarbageCollectorStats> entry : virtualMachineMetrics.garbageCollectors().entrySet()) {
+            final String gcName = entry.getKey().replace(" ", "").replace("-","_");
+            names.add((MetricNames.JVM_GC_RUNS + "_" + gcName).replace(".", "_"));
+            names.add((MetricNames.JVM_GC_TIME + "_" + gcName).replace(".", "_"));
+            types.add(typeFactory.createJavaType(long.class));
+            types.add(typeFactory.createJavaType(long.class));
+        }
 
         relDataType = typeFactory.createStructType(Pair.zip(names, types));
         return relDataType;
